Moved files to their correct place
