--- App.xaml
@@ -1,11 +1,0 @@
-<Application
-    x:Class="CourseApp.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:conv="using:CourseApp.Converters">
-    <Application.Resources>
-        <conv:InverseBooleanConverter x:Key="InverseBooleanConverter"/>
-        <conv:InverseBooleanToVisibilityConverter x:Key="InverseBooleanToVisibilityConverter"/>
-        <conv:AvailabilityColorConverter x:Key="AvailabilityColorConverter"/>
-    </Application.Resources>
-</Application>
