+++ CourseApp/Converters/AvailabilityColorConverter.cs
