+++ CourseApp/Views/ModulePage.xaml
@@ -1,4 +1,4 @@
-<Page
+﻿<Page
     x:Class="CourseApp.Views.ModulePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -9,7 +9,7 @@
             <RowDefinition Height="Auto"/>
             <RowDefinition Height="*"/>
         </Grid.RowDefinitions>
-        
+
         <!-- Header (Fixed) -->
         <Grid Grid.Row="0">
             <!-- Timer Display -->
@@ -20,7 +20,7 @@
                     <TextBlock Text="{Binding TimeSpent}" FontSize="16" Foreground="#E0E0E0" VerticalAlignment="Center" FontWeight="SemiBold"/>
                 </StackPanel>
             </Border>
-            
+
             <!-- Coin Container -->
             <Border Background="#2A2A2A" CornerRadius="20" Padding="16,8" HorizontalAlignment="Right" VerticalAlignment="Top"
                     BorderBrush="#454545" BorderThickness="1" Margin="0,0,0,20">
@@ -33,6 +33,16 @@
             <Button Content="Back" Click="BackButton_Click" Margin="0,0,16,0" Background="#2A2A2A" Foreground="#FAFAFA"  Padding="16,10" CornerRadius="8" BorderThickness="0"/>
         </Grid>
         
+        <Button Content="Complete Module" Command="{Binding CompleteModuleCommand}"
+        IsEnabled="{Binding IsCompleted, Converter={StaticResource InverseBooleanConverter}}"
+        Background="{Binding IsModuleAvailable, Converter={StaticResource AvailabilityColorConverter}}"
+        Foreground="White" HorizontalAlignment="Right"
+        Padding="16,10" CornerRadius="8" BorderThickness="0" Margin="0,0,0,20"/>
+
+        <!-- Add message when module is not available -->
+        <TextBlock Text="You must complete the previous module first."
+           Visibility="{Binding IsModuleAvailable, Converter={StaticResource InverseBooleanToVisibilityConverter}}"
+           Foreground="#FF6666" HorizontalAlignment="Right" Margin="0,5,0,20"/>
         <!-- Scrollable Content -->
         <ScrollViewer Grid.Row="1" VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Disabled">
             <StackPanel>
@@ -41,13 +51,13 @@
 
                 <TextBlock Text="{Binding CurrentModule.Description}" TextWrapping="Wrap" Margin="0,0,0,24" 
                        Foreground="#DADADA" FontSize="16" LineHeight="24"/>
-                
+
                 <!-- Image section -->
                 <Button Command="{Binding ModuleImageClick}" HorizontalAlignment="Center" Background="Transparent" 
                     BorderThickness="0" Padding="0" Margin="0,0,0,24">
                     <Image Source="{Binding CurrentModule.ImageUrl}" MaxWidth="1000"/>
                 </Button>
-                
+
                 <Button Content="Complete Module" Command="{Binding CompleteModuleCommand}"
                     IsEnabled="{Binding IsCompleted, Converter={StaticResource InverseBooleanConverter}}"
                     Background="#4F4FB0" Foreground="White" HorizontalAlignment="Right"
