--- CoursePage.xaml
@@ -1,112 +1,0 @@
-﻿<Page
-    x:Class="CourseApp.Views.CoursePage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:vm="using:CourseApp.ViewModels"
-    Background="#121212">
-
-    <Grid>
-        <Border x:Name="NotificationBanner"
-                Background="#4CAF50"
-                Padding="16"
-                CornerRadius="8"
-                HorizontalAlignment="Center"
-                VerticalAlignment="Top"
-                Margin="0,20,0,0"
-                Visibility="{Binding ShowNotification, Mode=OneWay}">
-            <TextBlock Text="{Binding NotificationMessage, Mode=OneWay}"
-                       Foreground="White"
-                       TextWrapping="Wrap"
-                       FontWeight="SemiBold" />
-        </Border>
-
-        <ScrollViewer VerticalScrollBarVisibility="Auto">
-            <Grid Margin="20">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto"/>
-                    <RowDefinition Height="*"/>
-                </Grid.RowDefinitions>
-
-                <!-- Header Section -->
-                <StackPanel Grid.Row="0">
-                    <Grid>
-                        <Grid>
-                            <!-- Timer Display -->
-                            <Border Background="#2D2D2D" CornerRadius="10" Padding="12,6" HorizontalAlignment="Left" 
-                                VerticalAlignment="Top" BorderBrush="#3F3F3F" BorderThickness="1">
-                                <StackPanel Orientation="Horizontal" Spacing="5">
-                                    <TextBlock Text="⏱️" FontSize="16" Foreground="#FFD700" VerticalAlignment="Center"/>
-                                    <TextBlock Text="{Binding TimeSpent}" FontSize="16" Foreground="#E0E0E0" VerticalAlignment="Center" FontWeight="SemiBold"/>
-                                </StackPanel>
-                            </Border>
-
-                            <Border Background="#2D2D2D" CornerRadius="10" Padding="12,6" HorizontalAlignment="Right" 
-                                VerticalAlignment="Top" BorderBrush="#3F3F3F" BorderThickness="1">
-                                <StackPanel Orientation="Horizontal" Spacing="5">
-                                    <TextBlock Text="{Binding CoinBalance}" FontSize="16" Foreground="#E0E0E0" VerticalAlignment="Center" FontWeight="SemiBold"/>
-                                    <TextBlock Text="🪙" FontSize="16" Foreground="#FFD700" VerticalAlignment="Center"/>
-                                </StackPanel>
-                            </Border>
-                            <!-- Progress Display -->
-                            <StackPanel Orientation="Horizontal" Margin="0,10,0,15" HorizontalAlignment="Center">
-                                <TextBlock Text="Progress: " FontSize="16" Foreground="#E0E0E0"/>
-                                <TextBlock Text="{Binding CompletedModules}" FontSize="16" Foreground="#9179FB" FontWeight="Bold"/>
-                                <TextBlock Text=" / " FontSize="16" Foreground="#E0E0E0"/>
-                                <TextBlock Text="{Binding RequiredModules}" FontSize="16" Foreground="#E0E0E0"/>
-                                <TextBlock Text=" modules completed" FontSize="16" Foreground="#E0E0E0"/>
-                            </StackPanel>
-
-                            <!-- Countdown Timer -->
-                            <StackPanel Orientation="Horizontal" Margin="0,0,0,15" HorizontalAlignment="Center">
-                                <TextBlock Text="Time remaining: " FontSize="16" Foreground="#E0E0E0"/>
-                                <TextBlock Text="{Binding TimeSpent}" FontSize="16" Foreground="#FFD700" FontWeight="Bold"/>
-                            </StackPanel>
-                        </Grid>
-                    </Grid>
-                    <TextBlock Text="{Binding CurrentCourse.Title}" FontSize="28" FontWeight="Bold" Margin="0,20,0,15" Foreground="#FFFFFF"/>
-                    <TextBlock Text="{Binding CurrentCourse.Description}" TextWrapping="Wrap" Margin="0,0,0,25" Foreground="#B0B0B0"/>
-
-                    <StackPanel Orientation="Horizontal" Margin="0,5,0,20">
-                        <Button Content="Back" Click="BackButton_Click" Margin="0,0,15,0" 
-                                Background="#2D2D2D" Foreground="#FFFFFF" Padding="15,8" CornerRadius="5"/>
-                        <Button Content="Enroll" Command="{Binding EnrollCommand}"
-                                IsEnabled="{Binding IsEnrolled, Converter={StaticResource InverseBooleanConverter}}"
-                                Background="#9179FB" Foreground="#FFFFFF" Padding="15,8" CornerRadius="5"/>
-                    </StackPanel>
-
-                    <Border Height="1" Background="#3F3F3F" Margin="0,0,0,20" Opacity="0.6"/>
-                </StackPanel>
-
-                <!-- Modules Section -->
-                <Grid Grid.Row="1">
-                    <StackPanel>
-                        <TextBlock Text="Modules:" FontSize="20" FontWeight="Bold" Margin="0,0,0,15" Foreground="#FFFFFF"/>
-
-                        <ListView ItemsSource="{Binding Modules}" x:Name="ModulesListView" 
-                                  IsItemClickEnabled="True" Background="Transparent" Margin="0,5,0,0">
-                            <ListView.ItemContainerStyle>
-                                <Style TargetType="ListViewItem">
-                                    <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
-                                    <Setter Property="Padding" Value="0,5"/>
-                                    <Setter Property="Margin" Value="0,5"/>
-                                </Style>
-                            </ListView.ItemContainerStyle>
-                            <ListView.ItemTemplate>
-                                <DataTemplate>
-                                    <Border Background="#1E1E1E" BorderBrush="#3F3F3F" BorderThickness="1" 
-                                            CornerRadius="8" Padding="15,12" Margin="0,3">
-                                        <StackPanel Orientation="Vertical">
-                                            <TextBlock Text="{Binding Title}" FontWeight="SemiBold" Foreground="#FFFFFF" 
-                                                       Margin="0,0,0,5" FontSize="16"/>
-                                            <TextBlock Text="{Binding ShortDescription}" TextWrapping="Wrap" Foreground="#B0B0B0" FontSize="14"/>
-                                        </StackPanel>
-                                    </Border>
-                                </DataTemplate>
-                            </ListView.ItemTemplate>
-                        </ListView>
-                    </StackPanel>
-                </Grid>
-            </Grid>
-        </ScrollViewer>
-    </Grid>
-</Page>
